Expand branding (#343)
note: kc-theme folder is a copy of origin
[keycloak-theme/apolo/login](https://github.com/neuro-inc/launchpad/tree/master/keycloak-theme/apolo/login)
folder with adjustments for branding.

The reason for copying is because we did not properly support versioning
[here](https://github.com/neuro-inc/launchpad/blob/master/charts/launchpad/values.yaml#L175)
and in-place update will break existing launchpad versions.

+++ launchpad/apps/service.py
@@ -873,8 +873,9 @@ class AppService:
             input=import_request.input,
         )
 
-    async def delete(self, app_id: UUID) -> None:
-        await self._apps_api_client.delete_app(app_id)
+    async def delete(self, app_id: UUID, uninstall: bool = False) -> None:
+        if uninstall:
+            await self._apps_api_client.delete_app(app_id)
         async with self._db() as db:
             async with db.begin():
                 await delete_app(db, app_id)
